move ScriptTester to its own module
