+++ cli/test-utils/src/lib.rs
@@ -5,6 +5,9 @@ mod macros;
 
 // Utilities for making it easier to handle tests.
 pub mod util;
-pub use util::{Retry, ScriptTester, TestCommand, TestProject};
+pub use util::{Retry, TestCommand, TestProject};
+
+pub mod script;
+pub use script::ScriptTester;
 
 pub use ethers_solc;
